pyproject.toml
[project]
name = "lab13-observability"
version = "0.1.0"
description = "Day 13 Observability Lab with comprehensive Gradio UI for monitoring, testing, and managing observability systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.118.0",
    "httpx==0.28.1",
    "langfuse==2.57.0",
    "prometheus-client==0.21.1",
    "pydantic==2.11.4",
    "pytest==8.3.5",
    "python-dotenv==1.1.0",
    "structlog==25.4.0",
    "uvicorn[standard]==0.37.0",
    "gradio==4.44.0",
    "huggingface_hub==0.25.2",
    "plotly==5.17.0",
    "pandas==2.1.4",
    "psutil==5.9.6",
]
